scraper: Modularise the config.
This is a change mostly intended because someone wanted to use it
and I wanted to provide an easier way to use it without them wasting
too much of their time in order to understand the logic behind
this. Comments have been added as such.

This might also help later on because I'm planning on making a CLI
for this, perhaps convert it into a runnable script? CLI is
definitely in the plan, the other is unsure.